Fix method of filling uncertain published relations
When we do not delete empty uncertain relation entries, pairwise relation accuracy is incorrect because algorithm and published relation dicts will not exactly match. For example, if we fill an uncertain 2nd-degree relation but do not delete the empty relation entry, we would have an extra {"2": 0} entry in the published relations dict.